[Design] #24 - 파일 정리
